common function created for download invoice
